start to use RenderInstance in an api-neutral fashion
